add `Map.remove` and related, adjust `Mutable...` emthods
There's now `Map.remove` (returns a `Map`) and `MutableMap.delete`
(returns `#void`), as well as `Set.remove` and `Mutableset.delete`.

For sets, the methods `union`, `intersect`, and `intersect` now exist
only on `Set`s. The old behavior was that the methods existed on
`ReadableSet`s but reported a contract error when called for a
`MutableSet`.