scribble/rhombus: rewrite spacer API and implementations
The spacer interface is the way that names in example code in
documentation get linked to the right definition in the right space
(expression vs. binding, for example). Each form that has
subexpressions different from itself needs a spacer that roughly
duplicates parsing, but the parse usually can be very coarse, since
the spacer only needs to identify which subcomponents are in what
space. The first cut was implemented in Racket, but the new version is
in Rhombus.