scribble/rhombus: change declared nonterminals
Change the way a nonterminal declared with `nonterinal` behaves.
Instead of changing the way the nonterminal name typesets in all
contests, it only affects contexts that refer to the name via `$` in
`doc` or that use `nontermref` explicitly.

This change should avoid many conflicts, such as `expr` being a
nonterminal as well as a prefix for things like `expr.macro`.