change stragegy for `use_static` and `use_dynamic`
Instead of defining multiple identifiers, like `.` and `for` and `++`,
just define `#%dynamism` and have all forms that want to be sensitive
to static mode check the binding of` #%dynamism` (through
`syntax_meta.is_static` in the case of Rhombus-implemented macros).
This way, the set of static-mode forms is naturally extensible, and we
avoid various problems redefining operators like `.`.

On a first impression, this might seem like a job for a syntax
parameter. But when a macro uses `.` in its expansion, I think the
macro will normally intend to use a static `.` or a dynamic `.`
depending on its own bindings (i.e., the definition site), not
depending on the bindings at the use site.