`Array.of`, `Array.now_of`, `Box.of`, and `Box.now_of`
The `Array.of` and `Box.of` annotations formerly only checked initial
content but still propagated static information for the content. Now,
`Array.of` is a converter annotation that is implemented by an
impersonator contract, while `Array.now_of` is a predicate annotation
that does not propagate static information for content, and the same
for `Box.of` and `Box.now_of`.