reduce bogus "expected more terms" errors
When a pattern along the lines of '$a ...: <body>' failed to match due
to the <body> not matching for some input, the error tended to be
"expected more terms", even though it's not possible to add a matching
': <body>' after a non-matching block. Avoid that misleadining error
by adding a constraint on '$a' that it doesn't match a block. The
constraint is added only when it can be easily inferred, so this is
not a complete solution, but it should avoid bad errors in common
cases.