revise `pattern` shorthand
Now that we allow a block before alts, the `pattern` syntax class
clause is redundant and arguably disfavored.  Not only that, it also
ties the meaning of the corresponding (unquote) binding operator to
it.  I think it'd be better that the syntax class clause simply be
removed, and the (unquote) binding operator be changed into an actual
shorthand for an inline syntax class matching.

My experience is that inline syntax classes are quite useful for the
same reason complex patterns in `syntax/parse` are useful, but using
them unfortunately requires an unnaturally large amount of rightward
shift.  Moreover, supplying `open` with inline syntax classes is
common but, again, looks ugly.  Therefore I believe `pattern` can
serve a better purpose this way.

Interestingly, syntax class clause macros originally can supply
spliced patterns through `pattern`.  This use is kept through changing
the expected expansion to a block optionally followed by alts.

A remaining problem is that the handling of repetitions *above* the
quasiquote doesn't seem to cover attributes.  Although this is not
directly related to `pattern` per se, the issue may be apparent when
the `pattern` binding is used in a composite binding form.  Again, the
fix should be generalizable.