use chaperones for `{Array,Box}.later_of`
This commit changes `{Array,Box}.later_of` to use chaperones for the
predicate case, and only use impersonators for the binding case.  The
binding case is changed to *not* discard the converted value.
Moreover, the error message is improved by including the annotation
and, for arrays, the position.

A bug was found that the binding case of `of` annotations did not
check the predicate before continuing.  This was fixed and some test
cases were added.