fix an innocent mistake in dot dispatch
The `call-with-values` in the assignment case appears to be wrong,
because the values produced by `build-assign` should be received by
`success-k`, not the other way around.  It *innocently* worked,
however, because `success-k` was just `values`, which produced, well,
no values, and then `build-assign` just returned the two values
without going through `success-k`.  Fix this innocent mistake by using
a `success-k` that does check for the correct arity, for future-proof
sake.