fix result contract in primitive `Listable.to_list`
More precisely, when the primitive functions are used directly.  In
such case, the `who` can be `#f`, but `raise-result-error*` must still
be called with `Listable.to_list` as the `who` (also, even when `who`
is provided, using it for the result contract is wrong).  In addition,
early quit with `#f` instead of raising when the value is not a
`Listable` but the `who` is `#f`.