add `Pict.children` and `Pict.metadata`
Add a way to get children of a pict, analogous to `pict-children` and
the Racket level. The `Pict.metadata` method is an alternative to
allowing subclasses of `Pict` to add new fields, and it could be
useful to implementing something like `ppict`. Also, add `~dhoriz` and
`~dvert` arguments to `Find`, inspired by `ppict` options. Meanwhile,
repair problems finding animated picts.