allow a static info key to have specific union and intersect operations
Default operations merge the same as `#%call-result`, and that default
is useful for the shortcut of using selector identifiers as
static-info keys for the selector's result. Otherwise, though, a
static info key should be bound to a compile-time record that supplies
union and intersection operations on values associated with the key.