Provide key-equality variants of maps via `Map.by`
In the same way that the `Map` constructor creates a map that uses
`==` and compatible hashing for keys, a new `Map.by(<key_comp>)`
constructor creates a map that uses a comparison operation and
compatible hashing function as bound to `<key_comp>`. Operators
`<key_comp>` are bound in a new space, and the predefined
`<key_comps>`s have the same names as comparison operators: `==`
(which is the default), `===`, `is_now`, and
`is_same_number_or_object` (which is `eqv?`).

Examples:
```
  Map.by(===){ make(): "a", make(): "b" }
  Set.by(===){ make(), make() }
```

The `Map.by` form can be used in the same places as `Map`, including
in binding patterns and as an annotation. A value constructed with
`Map.by(<key_comp>)` satisifies the `Map` annotation as well as
`Map.by(<key_comp>)`.

Ditto for `MutableMap`, `Set`, and `MutableSet`.

New key comparisons are bound with `key_comp.def` to implement custom
hasing for a map. That functionality requires a small addition to
Racket in the form of `unsafe-impersonate-hash`.