add `Comparable`
Change `<`, `<=`, `>`, and `>=` to be operations on comparable values
(instead of just numbers), and add `compares_equal` and
`compares_unequal` (which different from `==` and `!=` by going
through the comparable protocol instead of the equatable protocol).
Real numbers, characters, strings, byte strings, symbols, and keywords
are comparable.

The only required method of the `Comparable` interface is
`compare_to`, which should produce an integer that is negative for
less-than, 0 for equal-to, or positive for greater-than. Methods
`less`, `less_or_equal`, `greater`, `greater_or_equal`,
`compares_equal`, and `compares_unequal` can be overriden but use
`compare_to` by default.

Static information now requires that operations like `<` on
comparables can be statically resolved to a specific comparison
operation. Unlike `++`, either argument to `<` can determine the
specialization. If both arguments provide a specialization through
static information, the two must be consistent.

Add/keep `.<`, `.<=`, `.>=`, `.>`, and `.!=` as number-specific
operators. These dot-prefixed operators can be useful when it's not
convenient to have an annotation on one of the arguments. They might
also help in dynamic mode to be sure you're getting a number
operation.

The `Number` annotation specifies real-number comparisons, even though
not all numbers are comparable (i.e., complex numbers are not all real
numbers). Attaching only to annotations like `Real` would be more
precise, but then results for operations like `+` would be much more
complex to track usefully.

To support case-insensitive string and character comparsions, the new
veneers `CharCI`, `StringCI`, and `ReadableStringCI` are like `Char`,
`String` and `ReadableString`, but with case-insiensitive comparison.