extend `statinfo_meta.dot_provider_key` to support subclasses
Change the use of `statinfo_meta.dot_provider_key` values to allow a
sequence of identifiers, instead of just one identifier, where the
first identifier of a sequence is used, but intersection on sequences
finds a shared tail. That encoding allows an `if` with branches that
instantiate different classes to get dot information for a shared
superclass.

To keep things simple, this works only for single-inheritance cases
--- so superclasses and supervaneers, but not superinterfaces.