add `enum`
As an example,

```
HorizAlignment:
  left
  center
  right
```

defines `HorizAlignment` as an annotaton that is satisfied by
`#'left`, `#'center`, and `#'right`, and it also defines
`HorizAlignment.left`, `HorizAlignment.center`, and
`HorizAlignment.right` as expressions bound to those symbols and as
binding forms that match those symbols.

Sometimes, enumerated symbols are combined with other datatypes.
Although a union of an enumerated datatype with other annotations
could cover that case, `enum` provides direct support through `~is_a`
clauses that add other annotations directly to the enumeration. Here's
an example:

```
enum LineWidth:
  ~is_a Real
  inherit
```

The `pict` and `slideshow` libraries now use `enum` for various
enumerations instead of a hand-rolled annotation.