fix list `at_stx` argument in `syntax_meta.error`
Also correct the doc.  There was a naive correction in fe09a3e, but it
wasn't correct.  The extra `Syntax || List.of(Syntax)` argument, if
supplied, should correspond to `sub-expr` (and `extra-sources`) in
`raise-syntax-error`, while the other `Syntax` argument always
corresponds to `expr`.  That means the `Syntax` argument should be
`in_stx` ("in: ...." in error message), not `at_stx` ("at: ....").