anonymous-function shorthand with `_` and parentheses
When the terms enclosed by an expression `()` include an immediate
`_`, the `()` form (more precisely, the `#%parens` form) expands to an
anonymous function, and each `_` is a distinct argument. So, for
example `(_ + _)` is a function that takes two arguments and adds
them.

When a function-call expression form has an argument that is just `_`,
then the function call (more precisely, the #%call form) expands to an
anonymous function that contains the call, and each argument that is
just `_` becomes an argument to that enclosing function. So,
`List.cons(_, [1, 2, 3])` is a function equivalent to
`fun (x): List.cons(x, [1, 2, 3])`. This conversion does not apply
to a function call that has spliced arguments or `...`s.