add `WeakMutableMap` and `WeakMutableSet`
A weak mutable map is an ephemeron map at the Racket level. Although
there are uses that need a non-epheneron weak hash table, they're
pretty low-level and related to memory management. Meanwhile, Racket
CS's ephemeron-hash performance is about the same as for a weak hash
table. So, there's doesn't appear to be a need for non-ephemeron weak
maps.

A weak set uses a non-ephemeron hash table, though, since it's
equivalent to an ephemeron hash table for boolean values, and it's a
little simpler for the GC to manage.