implement integer range objects
This follows the proposal in "design/range/range.md".  Range objects
can be constructed and matched by nine constructors (but no separate
annotation is given for each constructor, to avoid the combinatorial
explosion).  The name `Range.from_to` is chosen for the common case
`..`, and other names build on it.  Four operators `..`, `..=`,
`<..<` (this is unfortunate, because operators cannot end in `.`), and
`<..=` are available as shorthand notations for the nine constructors.

Range objects are currently restricted to integers, because it's not
apparent how to make them generic while still efficient in the context
of Rhombus.

`Range.from_to`, `Range.from_to_inclusive`, and `Range.from` can be
used as sequences, and a naive optimization strategy is applied to
ensure that they work as efficiently as before.  The strategy simply
looks for a direct application of the three constructors and dispatch
to special cases that work like `in-range`, `in-inclusive-range`, and
`in-naturals`.  The same strategy is also found in `Range.step_by`,
but only when it's invoked as a dot method (over a direct
application).

`Range.from_to` and `Range.from_to_inclusive` are `Listable`, which is
the same as before.

Accordingly, ranges are separated into three subtypes: `Range` (all
ranges), `SequenceRange` (ranges that can be used as sequences), and
`ListRange` (ranges that are `Listable`).

A list of methods is implemented, which is essentially borrowed from
Rebellion.  In addition, ranges can be deconstructed through
`Range.start`, `Range.end`, `Range.includes_start`, and
`Range.includes_end`, if one wants to.