delay constructon of some compile-time values
Consistently thunk constructon of static-info syntax objects, and
delay creation of some other structures (such as classes) that contain
syntax objects.

The goal of ths change is to reduce the time needed to visit `rhombus`
(in the sense of a phase-1 instantiation) and `rhombus`-based modules.
The savings turn out to be very small for just `rhombus`, though.