repairs for dotted-name resolution
The search to resolved a dotted-name binding was incorrectly
conflating identifiers for a space-specific binding and for a
namespace binding. Exporting a name in the `namespace` space as well
as a non-expression space was one way to expose the problem. There
were also related problems with re-exporting bindings via namespace
paths.