allow `field` right-hand sides to refer to earlier fields
For example,

 class Posn(x, y):
   field dist = math.sqrt(x**2 + y**2)

initialized the `dist` field in the way you'd expect. Note that
`this`, methods, or later fields are not available. This makes the
binding environment of the right-hand side of `field` somewhat
unusual, but perhaps no more unusual that they way methods can refer
to `this`, etc.

Making this work triggered a substianl cleanup in the internal
implementation of constructors.