fix `&` splicing of mixed-kind maps/sets
Now that we have map configurations, the assumption in `build-setmap`
was too strong.  It assumed that the first element, if a spliced
element, just needs to be asserted as a map/set, which was correct
when we didn't have map configurations.  Instead, we must generally go
through the append operations, which only optimize for cases when we
have same-kind maps/sets.

Also fix `...` splicing, both the optimization and wrong behavior in
`Map`s.

As an example, use this kind of splicing in `draw.Font`.