change repetition handling to delay expressions
Make examples like these work better:

```
> let [x, ...] = [1, 2, 3]
> [(x == 1) && println(x), ...]
1
[#void, #false, #false]
> [(x != 1) && println(x), ...]
2
3
[#false, #void, #void]
```

In particular, an expression derived from a use of a repeition is
evaluated once for each instance of the repetition. In contrast, the
old implementation would create a list of `x == 1` results and
separate list of `println(x)` results, and then build up a larger
klist from those.

The documentation still specify repetitions well enough to make this
distinction, and the implementation creates too many intermediate
sequences. So, this is just a first step.