change representation of repetitions
Instead of an expression that produces nested pair lists, change the
representation of a repetition to be a sequence of `for` clauses plus
a `for` body. This revised representation is a better fit for the
revised semantics of repetitions, it makes repetition combinations
much simpler to implement, and it produced faster code by deforesting
common cases.

This changes requires some small adjustments to the binding macro
protocol and the repetition macro protocol. The current repetition
macro protocol is still tied to lists (but regular Rhombus lists
instead of pair lists), and we'll want to add a packing variant to
`repet_macro` to support for more general forms. Related to that, the
binding macro protocol involves "sequencer" forms that are currently
opaque. Adding a way to go from an annotation to a sequencer will fill
in the binding protocol and probably be the right ingredient for a
generalized repetition packer.