don't discard syntax props added by enforestation
When we unwrap/discard static-info wrappers on sub-expressions,
transfer all syntax props to the result, which can include those added
by enforestation (in particular, syntax-origin ones).  This solves the
problem such that `1 + 2 - 3` only includes the syntax-origin prop for
`-`.

Also flip the order in which enforestation applies origin-tracking and
scope-flipping, following the order in the Racket expander.