enable repetition-sensitive dot providers
When a dot form needs extra arguments, then it's not enough to handle
the left-hand side as a repetition automatcially if the arguments can
also be repetitions. Keep automatic handling for the simple case, and
allow a dot provider to be called in repetition mode to handle
repetition arguments. A dot handler opts into repetition support by
including the `~is_repet: var` option.

Closes #389