pict: add support for time-parameterized finders
Along with the other recent improvement to `pin`, this change makes it
much easier to add common types of animations to an existing animation.

To enable time-parameterized finders, `animate_map` threads through an
epoch and time within the epoch. So, that enabled time-sensitive
lifting in general.

This commit also changes the `~duration` argument of `pin` and
`balloon` back to `DurationAlign` and adds a separate `~time` argument
as a `TimeAlign`, since insertion of epochs is orthogonal to padding
versus sustaining. This change also makes time-based alignment of the
pinned pict separate from the decision to insert epochs or not,
although those are still bunled into the single `~time` argument.