Add `pict/rhombus` to render literal shrubbery forms to picts
The `rhombus` and `rhombusblock` forms of Scribble and pict work the
same, except that the results are for different targets, and escapes
must produce different kinds of values. The two share a common core
that's in a new "shrubbery-render" package.

The shared core means that `rhombus/pict` could use `meta_label`
binding information to style its output, but that isn't set up, yet.
Also, the documentation for "shrubbery-render" is just a sketch.