shrubbery: add raw inner prefix and suffix properties
I'm not too happy with further complication in raw text properties,
but I don't have a better idea right now for making `@` notation stick
with its term for documentation examples (e.g., examples for `str`).

A lot of the commit is about improving the documentation for
representating shrubbery forms with raw text properties. A cleaner and
clearer separation of Racket and Rhombus notation is still needed.