shrubbery: make `\` handling more normal
The original `\` rules were trying too hard to enable something that
hasn't turned out to matter: being able to reach any virtual column
within a narrow physical column. The attempt made the interaction of
`\` and indentation very difficult to parse visually.

Make `\` much simpler by not affecting the way that columns are
assigned to tokens on the next line. If a continuing line has an
opening parenthesis or `:` followed by another token, for example,
that token's column (potentially needed by subsequent lines) is the
visually obvious one, instead of a column that depends on the
continued line's length. Editor indentation support encourages lining
up a continuing line's start with the group's start; something more
specialized would be possible, but that's the simplest and most
obvious choice, and it looks reasonable.

Also, disallow `\` followed by more on the same line. Previously, a
`\` with mroe on the same line that was treated as whitespace, but
whitespace `\`s haven't turned out to be useful, either. It's less
confusing to disallow them. Disalloweing them also potentially leaves
a little more room for future uses of `\`; then again, `\` needs to be
treated as whitespace after all in line-insensitive mode, so that may
limit future meanings for `\`.