Scribble: support module-prefixed documentation entries
Support documenting a form using a module-qualified name. For example,
`def` from the `racket/runtime_path` module is documented as
`runtime_path.def`, since it's meant to be used that way, and the name
`def` by itself is confusing.

When a binding is documented with a module prefix, the prefix shows up
non-bold at the definition site.