add `->` and `Function.all_of` annotation forms
The `->` annotation operator constructs a converting annotation that
expects a function of suitable arity and wraps it to check argument
and result annotations. The `Function.all_of` annotation form takes
any number of annotations constructed by `->` and creates a annotation
that is satified by a function that handles all of `->` cases, similar
to `case->` in Racket's contract system.

The expressiveness of `->` itself is similar to `->i` in Racket's
contract system: it supports optional and keyword arguments, and
arguments can be named to create dependent checks.

Unlike `case->`, which selects a case based only on the arity of a
call, `Function.all_of` selects a case based on the first clause that
matches the arguments of a given call. This makes `Function.all_of`
more analogous to a multi-case `fun`, and it means that
`Function.all_of` can express some dependencies even with simple `->`
forms, as in

```
  Function.all_of(Int -> Int,
                  String -> String)
```

To get error reprting to work well overall, this commit makes a
general change to the annotation-macro layer. Each annotation form is
now responsible for attaching its rendered form to its expansion, like
expression macros. So, there are more internal uses of
`relocate+reraw`, and there's more need for `Syntax.relocate_span`s at
the Rhombus level.

The `Any` annotation is like `any/c` in Racket, which means that
`() -> Any` will wrap a function with a non-tail call to ensure that it
produces a single result. Use `() -> (Any, ...)` or `() -> (& Any)` to
avoid constraining the number of results and produce a tail-calling
wrapper.

The `->` operator cannot by itself handle multiple arguments in
parentheses on the left of the operator, since an infix operator takes
a parsed left-hand side, and a parenthesized sequence of annotation is
not an annotation. As a workaround, the `#%parens` form exported by
`rhombus` cooperates with the `->` form.