implement a conservative optimization in `Function.pass`
With this change, a simple `Function.pass(arg, ...)` call is optimized
so that no intermediate lists are produced, much like a `for` loop.
The call is "simple" if every `arg` is either a simple expression or
repetition, so we skip the optimization if there is any keyword
argument or rest argument.

This optimization is by and large inspired by the use of
`Function.pass` in Shplait's internal.