reference: move away from `dispatch_table`
Indicate the distinction between methods and non-methods at the point
of documenting a function, instead of in a separate table at the start
of the section. To support more flexible methods (e.g., a `String`
method that applies to `ReadableString`s, also), genralize the
`method` documentation form. A new section of the manual explains
documentation conventions.