namespace prefixes on inferred names, and `~name` for `class`
When a `fun` forms is used inside a `namespace` body, the default
run-time name of the function is now prefixed by the namespace. This
change was motivated by moving the `filesystem` namespace's
implementation from Racket to Rhombus.

Along similar lines, methods are intended to have run-time names
prefixed by their class. That now mostly works, but it will probably
take time to track down all the right places that need to adapt.

The `class` form now supports `~name` analogous to `fun`. When `~name`
is used in `fun` or `class`, it does not automatically get a prefix
from an enclosing namespace.

To communicate a namespace prefix, the definition-macro protocol gets
a new argument, which is exposed at the Rhombus level through a
`~name_prefix` option in `defn.macro`, `decl.nestable_macro`, and
`defn.sequence_macro`.