repair arity checking and messages for methods
For a method call through an object followed by `.`, make sure that
when the wrong numeber of argument is provided, the report includes
the expected number of arguments not counting the implicit `this`
argument. That's now implemented by simply marking the
method-implementing function as a method in the sense of Racket's
`procedure->method`. But, then, the same function cannot be used for
the namespaced-function variant of the same operation --- the one that
takes the target object as an explicit first argument. To keep
reporting right for the latter, create a new wrapper function that
accepts the same arguments as the method form and just passes them
along. If a method is accessed from an object without calling the
method, a curried form of the method is created dynamically.

Primitive "classes" like String continue to work in roughly the
opposite mode: a namespaced function like `List.delete` contains the
actual implementation, and a method-style `.delete` call is
transformed by a dot provider to call the namespaced function. If the
method-style call has statically the right number of arguments, it is
translated to a direct call. Otherwise, the conversion uses a
predefined curried version of the operation that bounces arguments to
the uncurried version.

The two strategies are equally good for static calls to known methods.
For dynamic or higher-order use of methods (i.e., finding a method
when the class is not statically known or getting a method without
calling it), the primitive-class strategy is better overall, but it
relies on the primitive "class" being final. The non-primitive
strategy is better when methods can be overridden. Many Rhombus
classes will turn out to be final, so there's room for improvement by
using the primitive-like stragegy for final classes, which would
better support dynamic contexts.