generalize list binding patterns to support multiple splices
Allow multiple `...` repetitions and `&` splices in a list pattern,
and also allow repetitions of splices. In general, matching such
patterns can require a search, and there can be multiple matches. This
generalization is the same as in syntax patterns, and the search
implemented here is similarly greedy (i.e., longer matches to earlier
subpatterns are preferred).

The low-level binding protocol was not quite right to support a
backtracking search, so much of this commit is adding another element
to the low-level protocol. Although making the protocol yet more
complex is worrying, supporting composable backtracking search seems
worthwhile. Also, the set of predefined forms is rich enough ---
especially with `&&`, `||` and `where` patterns --- to expect that
most binding macros are (and will be) derived from the predefined
ones, instead of written directly with the low-level protocol.

One difference between syntax matching and list matching is that
there's no notion of splicing bindings forms, whereas syntax matching
has splicing syntax classes. But the same effect can be achieved by
making the use of a "splicing" binding explicit with `&`. For example:

```
  bind.macro 'Two($p)':
    '$p && [_, _]'

  match [1, 2, 3, 4, 5, 6]
  | [& Two(p), ...]:
      [p, ...]
  // => [[1, 2], [3, 4], [5, 6]]
```

Splicing matches can even be efficient, because `[]` patterns generate
static information that describes list bounds. In the example above,
the pattern `$p && [_, _]` is statically known to match 2 elements, so
`[& Two(p), ...]` will only attempt to match 2-element sequences to
`Two(p)`. In contrast, `[& [a, ..., "x", b, ...], ...]` would take
quadratic time to fail to match a list that has no `"x"`, because it
would try all possible prefixes of an input list, and then each prefix
list would be searched (in linear time) for an "x".

Pair-list patterns support the same repetition and splicing forms as
list patterns, but matching a splice may be less efficient, because
constructing a candidate subsequence to match the splice takes time
O(N) for a subsequence of length (N).