scribble: method linking in `rhombus` and `rhombusblock`
Change the rendering pass used by `rhombus` and `rhombusblock` to
recognize syntax properties (intended to be added by spacers) to
support linking a method or property name that appears after a `.`.
Linking requires inferring an annotation-as-namespace for the term
before the `.`. When the term sequence before the `.` is a function
call, then result-annotation information needs to have been registered
as part of the documentation of the function, and so on.

Like the spacer approach in general, this approach is approximate, but
a little inference of this form goes a long way. Most identifiers
after `.` in current Rhombus documentation are now linked. The biggest
gap is propagating information from identifier bindings to uses.