generalize group positions in a syntax pattern
Instead of allowing only an escape at the end of a group to serve as a
group escape, allow any escape within a group that is not followed by
an escape to be a group position.

The intent is to allow something like

```
expr.macro 'with_options $(e :: expr_meta.Parsed):
              $options':
  ....
```

instead of requiring that to be written as

```
expr.macro 'with_options $term ...:
              $options':
  let '$(e :: expr_meta.Parsed)' = '$term ...'
  ....
```

This changs is not backward-compatible, because it can change the
meaning of existing syntax patterns. However, nothing in this repo (or
the Shplait repo) seems to be broken by the change. The revised rule
seems no more difficult to explain, and it seems in line with other
conventions in Rhombus to make patterns work as expected.