add `!!`, `?.`, and `?>`
The `!!` form is a destructor to go with the `maybe` constructor. As
an expression, `expr!!` produces the same result as `expr`, except
that it errors if that result is `#false`. And if `expr` has static
information from `maybe(annot)`, then `expr!!` has the static
information of `annot`. The `bind!!` binding form matches only
non-`#false` values that match `bind`. Using `!!` is the same as
writing `expr :: annot` or `bind :: annot`, except that you don't have
to write out `annot`.

Example:

```
use_static
// if literal syntax is known to have a source location
'stx'.srcloc()!!.line
// if not known
match 'stx'.srcloc()
| loc!!: loc.line
| ~else: #false
```

The `?.` is like `.` with the shape of a field access or method call.
If the expression to the left of `?.` produces `#false`, then result
is `#false`. If the expression to the left of `?.` is not `#false`,
then it is used as the left-hand expression for the `.`, and it gets
the static information of `annot` if `expr` has the static information
of `maybe(annot)`. The `?>` operator is analogous for `|>`.

```
use_static
'stx'.srcloc()?.line
'stx'.srcloc() ?> Srcloc.line
```