revise syntax-pattern matching context rules
The original goal was to make this example from @samdphillips work:

```
syntax_class Option:
  root_swap: o orig
| '~opt_a: $n':
    field o = {#'a: n}
| '~opt_b':
    field o = {#'b: #true}

syntax_class.together:
  syntax_class Options:
    fields: o
  | '$(opt :: Option)': field o = opt
  | '$(opt :: Option)
     $(tail :: Options)': // not a multi-group context, but should be
      field o = opt ++ tail.o

check:
  match '~opt_a: 10; ~opt_b'
  | '$(v :: Options)': v.o
  ~matches { #'a: '10; ~opt_b' }
```

Generalizing multi-group patterns in the obvious way to allow
multi-group escapes turns out to have bad interactions with the
existing escape rules. For example, it makes the following `$y` match
any number of groups (i.e., any number of extra arguments), even
though it looks like a pattern that matches two arguments:

```
 'f($x, $y)'
```

The solution in this commit is to revise the rules about multi-term
and multi-group sequences to make identifier escapes match sequences
*less* often, but to *more* often allow a syntax-class annotation to
opt into multi-group matches (as well as multi-term matches). The
"more often" part accomodates the `Options` example above. The "less
often" part is the backward-incompatible part; that change is not
strictly required to make the "more often" cases work, but changing
makes the rules simpler and more consistent.

The key difference is that flexible matching in a group or multi-group
context must be enabled by asking for it with a syntax class. An
identifier escape by itself will default to a simple match (and
arguably a more expected one). Having to specify a multi-term or
multi-group match resolves bad interactons between the two cases in a
consistent way.

Specifically, if an identifier escape is not alone with its group, it
no longer matches multiple terms, even if it's at the end of the
group. Formerly, the `$y` in the following pattern could match
multiple terms:

```
 '$x + $y' // will only match 3 terms
```

The former rule allowing multiple terms for `$y` was intended to
support group patterns like `$(y :: expr_meta.Parsed)`. That still
works; it's just that the default for an unannotated pattern variable
is a single-term match, and the generalization to match
`expr_meta.Parsed` is enabled by asking for it.

If an escape is alone within its group, it is still eligible to match
a whole group, because that part has worked well and is widely used.
For example, `$x` below still matches a whole group, and not just one
term:

```
 'f($x)'
```

Formerly, though, the `$x` above could match multiple groups, because
the group containing `$x` is also alone within its sequence of groups
in `()`. I found instances where that was not the intended behavior,
and macros could be broken by providng unexpected groups. With this
commit, the "alone within group sequence" rule is disabled for `()`,
`[]`, and `{}`, but it is kept for blocks and `''`:

```
 'thunk: $body' // still matches a multi-group body
```

Multi-group matching in a block or `''` has worked well and seems
widely used. A potential rationalization for the difference (or
explanation for current practice) is that groups are not separated by
commas in a block or `''`, but they are comma-separated in `()`, `[]`,
and `{}`.

A syntax class can be used to get back to the old behavior in
multi-term or multi-group cases:

```
 'f($(x :: Multi))'      // any number of "argument" groups
 '$x + $(y :: Sequence)' // matches 2 or more terms
 '$x + $(y :: Group)'    // 3 or more terms, since groups are nonempty
```

With the change, most existing Rhombus code works as before, but
changes are sometimes needed --- usually along the lines of adding
`:: Multi`, `:: Sequence`, `:: Group`, or changing a `$y` to `$y ...`.
Several macro and spacer implementations are updated in this commit;
it's a small number compared to the number of patterns in the code,
and it's maybe representative of the level of change needed in other
Rhombus programs.